fix: move Vector classes to subpackage
