+++ raytracer/src/main/java/rayrangers/raytracer/math/Vector4D.java
@@ -1,4 +1,4 @@
-package rayrangers.raytracer;
+package rayrangers.raytracer.math;
 
 /**
  * Represents a vector in 4D space.
